fix: minor style changes, renaming files to match consistent pattern, and removing unnecessary styles
