--- src/components/TeamMember/projectsynthesis.code-workspace
@@ -1,12 +1,0 @@
-{
-  "folders": [
-	{
-		"path": "../../.."
-	},
-	{
-		"name": "backend",
-		"path": "../../../../../synthesis_backend/backend"
-	}
-],
-  "settings": {}
-}
