+++ src/components/button/button.stories.tsx
@@ -4,7 +4,7 @@ import { Button } from "./button";
 
 const meta: Meta<typeof Button> = {
   component: Button,
-  title: "Button",
+  title: "Components/Button",
   tags: ["autodocs"],
   argTypes: {
     variant: {
