+++ src/components/team-member/index.ts
@@ -1,2 +1,1 @@
-export * from "./button";
 export * from "./team-member";
